# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DSA Studio API
  description: REST API for the DSA Studio learning platform — through Phase 4 Advanced Features.
  version: 0.4.0
  contact:
    name: DSA Studio Team
servers:
  - url: http://localhost:4000/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /topics/progress:
    get:
      tags: [Topics]
      summary: User completion per topic
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Topic progress breakdown
...
  /topics/{id}/questions:
    get:
      tags: [Topics]
      summary: List questions for a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Questions in topic
  /questions:
    get:
      tags: [Questions]
      summary: List questions with filters
      parameters:
        - name: topicId
          in: query
          schema:
            type: string
            format: uuid
        - name: topicSlug
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [basic, intermediate, advanced]
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated question list
  /questions/{id}:
    get:
      tags: [Questions]
      summary: Get question detail (sample test cases only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Question detail
        '404':
          description: Question not found
...
  /questions/{id}/hints:
    get:
      tags: [Questions]
      summary: Get progressive hints up to tier N
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tier
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Revealed hints
  /questions/{id}/solutions:
    get:
      tags: [Questions]
      summary: Get solutions (after first attempt)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Solution list
        '403':
          description: Solutions locked until first attempt
  /questions/{id}/starter:
    get:
      tags: [Questions]
      summary: Get starter code template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: language
          in: query
          schema:
            type: string
            enum: [python, javascript, typescript]
            default: python
      responses:
        '200':
          description: Starter code
  /questions/slug/{slug}:
    get:
      tags: [Questions]
      summary: Get question by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Question detail
...
  /progress:
    get:
      tags: [Progress]
      summary: Overall progress, streak summary, earned badges
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Progress overview
  /progress/daily:
    get:
      tags: [Progress]
      summary: Daily activity history
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated daily activity
    post:
      tags: [Progress]
      summary: Log or update daily activity manually
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                activityDate:
                  type: string
                  format: date
                questionsAttempted:
                  type: integer
                questionsSolved:
                  type: integer
                timeSpentMinutes:
                  type: integer
                topicsCovered:
                  type: array
                  items:
                    type: string
                notes:
                  type: string
      responses:
        '201':
          description: Activity logged
  /progress/streak:
    get:
      tags: [Progress]
      summary: Streak information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current and longest streak
  /progress/analytics:
    get:
      tags: [Progress]
      summary: Analytics data for charts and badges
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Heatmap, weekly bar, topic pie, badges
...
  /chat/query:
    post:
      tags: [Chat]
      summary: General DSA chat query
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Assistant reply with session id
  /chat/hint:
    post:
      tags: [Chat]
      summary: Progressive AI hint for current problem
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tiered hint without full solution
  /chat/review:
    post:
      tags: [Chat]
      summary: AI code review
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Correctness, complexity, and optimization feedback
  /chat/history:
    get:
      tags: [Chat]
      summary: Paginated chat session list
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Chat sessions with last message preview
  /chat/history/{sessionId}:
    get:
      tags: [Chat]
      summary: Messages for a chat session
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full session message history
  /chat/learning-path:
    get:
      tags: [Chat]
      summary: Personalized learning path recommendations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Rule-based and AI-suggested topic sequence